Fix where `Current` was not locking to the thread
`Current` is an exported function and needs to be locked to the thread
otherwise while itterating we may get moved to another thread and wind
up with mixed namespaces.

This fixes an areal issue in the code where we use `Current` unlocked.
This showed up in tests where it couldn't open the namespace file due to
"no such file or directory".

Signed-off-by: Brian Goff <cpuguy83@gmail.com>