Add some support for CLONE_NEWUSER
This uses a raw clone(2) call to fork a new process in order to create a
user namespace (and any other namespace that was specified along with
it).

Creating a Set with a user namespace will make it so that `Set` cannot
be used with `Do` unless `Dup` is called on it along with all the
desired flags *except* for CLONE_NEWUSER. This effectively creates a new
set with all the new namespaces except for the user namespace.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>